add initial refactor for 'index.ts'; add bare minimum typescript configs and dependencies
