+++ package.json
@@ -2,8 +2,9 @@
   "name": "ryderserial-proto",
   "version": "0.0.1",
   "description": "",
-  "main": "src/index.js",
+  "main": "out/index.js",
   "scripts": {
+    "build": "tsc -p .",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -11,6 +12,10 @@
     "url": "git+https://github.com/Light-Labs/ryderserial-proto.git"
   },
   "author": "Marvin Janssen",
+  "contributors": [
+    "Marvin Janssen",
+    "Chaz Kiker <chazkiker@gmail.com> (https://github.com/chazkiker2)"
+  ],
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/Light-Labs/ryderserial-proto/issues"
@@ -18,5 +23,11 @@
   "homepage": "https://github.com/Light-Labs/ryderserial-proto#readme",
   "dependencies": {
     "serialport": "^9.0.6"
+  },
+  "devDependencies": {
+    "@tsconfig/recommended": "^1.0.1",
+    "@types/node": "^15.3.0",
+    "@types/serialport": "^8.0.1",
+    "typescript": "^4.2.4"
   }
 }
