+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "@tsconfig/recommended/tsconfig.json",
+	"compilerOptions": {
+		"outDir": "./out",
+		"lib": [
+			"es6",
+		]
+	},
+	"include": [
+		"./src/**/*"
+	],
+	"exclude": [
+		"node_modules"
+	]
+}
